fix: min(), max() macros incompatible with ESP32 BLE library The original Uno platform allowed these macros to compare between dissimilar types. They were defined for ESP32 in the platform file to allow comparison between unsigned and signed types. This causes some strange issue with the BLE library. The definitions have now been removed, and the problem fixed at the source: parameters passed to min() and max() are now typecast before comparison.

+++ src/Displays/BaseDisplay/Bounds/window.cpp
@@ -72,7 +72,7 @@ uint16_t WindowBounds::getWindowBounds(WindowBounds::side request) {
 
     // Handle a special case; funny issues with unusual drawing_width
     if (rotated_request == R)
-        result = min(result, drawing_width - 1);
+        result = min(result, (uint16_t)(drawing_width - 1));
 
     // If required by LUT, find display width or height, and subtract the edge distance
     if (rotswap) {
